Extract specified agricultural community boundary data

extract_boundary() extracts the specified data from the list returned by get_boundary().

Usage

extract_boundary(boundary, city, old_village = "", community = "", all = FALSE)

Arguments

boundary

List of one or more agricultural community boundary data provided by the MAFF.

city

A local government name in Japanese to be extracted. In the case of overlapping local government names, this must contain the prefecture name in Japanese and the prefecture code in romaji (e.g., "Fuchu-shi, 13", "fuchu 13", "34 fuchu-shi", "34, FUCHU-CHO"). Alternatively, it could be a 6-digit local government code.

old_village

String by regular expression. One or more old village name in Japanese to be extracted.

community

String by regular expression. One or more agricultural community name in Japanese to be extracted.

all

logical.

Value

A list of sf::sf() object(s).

See also