Get the agricultural community boundary data

get_boundary() downloads and reads one or more agricultural community boundary data provided by the MAFF.

Usage

get_boundary(data, year = 2020, quiet = FALSE, path = NULL, to_wgs84 = TRUE)

Arguments

data

List of sf::sf() objects or one or more strings representing prefecture codes.

year

Year in which the agricultural community boundary data was created.

quiet

logical. Suppress information about the data to be read.

path

Path to the ZIP file containing the agricultural community boundary data; use a local ZIP file instead of going looking for a ZIP file. Specify a directory containing one or more ZIP files, not the ZIP file itself.

to_wgs84

logical. Convert JGD2000 to WGS 84.

Value

A list of sf::sf() objects.

Examples

if (FALSE) { # interactive()
path <- system.file("extdata", "castle.zip", package = "fude")
d <- read_fude(path)
b <- get_boundary(d)
}