Itemize the structure of Fude Polygon data

ls_fude() lists the year and the local government names (or codes) in order to understand what is included in the list returned by read_fude().

Usage

ls_fude(data)

Arguments

data

List of sf::sf() objects.

Value

A data.frame.

See also

Examples

path <- system.file("extdata", "castle.zip", package = "fude")
d <- read_fude(path, stringsAsFactors = FALSE, quiet = TRUE)
ls_fude(d)
#>    full_names year  names local_government_cd city_kanji        romaji
#> 1 2021_382019 2021 382019              382019     松山市 Matsuyama-shi
#> 2 2022_382019 2022 382019              382019     松山市 Matsuyama-shi
#> 3 2022_382078 2022 382078              382078     大洲市       Ozu-shi